feat: add keepalive secs

+++ crates/restsend/src/lib.rs
@@ -33,7 +33,7 @@ const MAX_CONNECT_INTERVAL_SECS: u64 = 5; // 5 seconds
 #[allow(unused)]
 const MAX_ATTACHMENT_CONCURRENT: usize = 12;
 
-const KEEPALIVE_INTERVAL_SECS: u64 = 50; // 50 seconds
+const KEEPALIVE_INTERVAL_SECS: u32 = 50; // 50 seconds
 const MEDIA_PROGRESS_INTERVAL: u128 = 300; // 300ms to update progress
 const CONVERSATION_CACHE_EXPIRE_SECS: i64 = 60; // 60 seconds
 const USER_CACHE_EXPIRE_SECS: i64 = 60; // 60 seconds
